requirements.txt
requests>=2.26.0
openai-whisper>=20240101
tqdm>=4.65.0
pytest>=7.0.0
pytest-mock>=3.10.0 